Remove unnecessary header
